backend/app/jwt_handler.py
from datetime import datetime,timedelta,timezone
from jose import JWTError,jwt
from . import models,database


SECRET_KEY="mysecretkey_please_change_in_production"
ALGORITHM="HS256"
ACCESS_TOKEN_EXPIRE_MINUTES=30

def create_access_token(data:dict):
    to_encode=data.copy()
    expire=datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp":expire})
    encode_jwt=jwt.encode(to_encode,SECRET_KEY,algorithm=ALGORITHM)
    return encode_jwt

def verify_token(token:str,credential_exception):
    try:
        payload=jwt.decode(token,SECRET_KEY,algorithms=[ALGORITHM])
        user_id:int=payload.get("sub")
        if user_id is None:
            raise credential_exception
        return user_id
    except JWTError:
        return credential_exception
